third_party/blink/web_tests/external/wpt/css/css-text-decor/parsing/text-decoration-inset-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property text-decoration-inset value '0px'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value '0'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value '0px 0px'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value '0.5em'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value '-1em'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value '-1ch -1ch'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value '1px 2px'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value '1ch -1ch'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value 'calc(1em / 4) calc(-1ch)'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
[FAIL] Property text-decoration-inset value 'auto'
  assert_true: text-decoration-inset doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
